requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
streamlit-chat>=0.1.1
streamlit-elements>=0.1.0
chromadb>=0.4.0
langchain>=0.2.14
langchain-community>=0.2.12
langchain-chroma>=0.1.4
langchain-huggingface>=0.0.3
python-dotenv>=1.0.0
numpy>=1.24.0
google-generativeai>=0.8.3
torch>=2.3.0; platform_system!="Windows" or platform_machine!="x86_64"
sentence-transformers>=2.7.0